+++ test/e2e/baremetal_e2e_test.go
@@ -18,6 +18,7 @@ package main
 
 import (
 	"flag"
+
 	"os"
 	"testing"
 
@@ -37,21 +38,17 @@ func skipIfNotCI(t *testing.T) {
 	}
 }
 
-func registerBMDriverFlags(flags *flag.FlagSet) {
-	flags.BoolVar(&common.BMDriverTestContext.BMDeploySchedulerExtender, "bm-deploy-scheduler-extender",
-		true, "Deploy extender for scheduler")
-	flags.BoolVar(&common.BMDriverTestContext.BMDeploySchedulerPatcher, "bm-deploy-scheduler-patcher",
-		true, "Deploy patcher for scheduler config")
-	flags.BoolVar(&common.BMDriverTestContext.BMWaitSchedulerRestart, "bm-wait-scheduler-restart",
-		true, "Wait for scheduler restart")
-	flags.BoolVar(&common.BMDriverTestContext.BMDeployCSIBMNodeOperator, "bm-deploy-csi-bm-node-operator",
-		true, "Deploy controller for CSIBMNode CRs")
+func registerCustomFlags2(flags *flag.FlagSet) {
+	flags.StringVar(&common.BMDriverTestContext.ChartsDir, "chartsDir",
+		"/tmp/charts", "Path to folder with helm charts")
+	flags.BoolVar(&common.BMDriverTestContext.CompleteUninstall, "completeUninstall",
+		true, "Uninstall pvc, volumes, lvgs, csibmnodes")
 }
 
 func init() {
 	framework.RegisterCommonFlags(flag.CommandLine)
 	framework.RegisterClusterFlags(flag.CommandLine)
-	registerBMDriverFlags(flag.CommandLine)
+	registerCustomFlags2(flag.CommandLine)
 }
 
 func Test(t *testing.T) {
